Make paperclip icon bigger and move heading links init code to separate plugin file.
