+++ site/public/styles/global.css
@@ -299,6 +299,10 @@ nav > ul {
   opacity: 0;
 }
 
+.anchor-icon svg {
+  vertical-align: bottom;
+}
+
 .anchor-icon:hover {
   opacity: 1;
 }
